+++ crates/jarl/tests/integration/comments.rs
@@ -29,7 +29,6 @@ any(is.na(x))
     success: true
     exit_code: 0
     ----- stdout -----
-
     ── Summary ──────────────────────────────────────
     All checks passed!
 
@@ -66,7 +65,6 @@ any(is.na(z))
     success: true
     exit_code: 0
     ----- stdout -----
-
     ── Summary ──────────────────────────────────────
     All checks passed!
 
@@ -212,7 +210,6 @@ x = any(is.na(y))
     success: true
     exit_code: 0
     ----- stdout -----
-
     ── Summary ──────────────────────────────────────
     All checks passed!
 
@@ -251,7 +248,6 @@ foo(
     success: true
     exit_code: 0
     ----- stdout -----
-
     ── Summary ──────────────────────────────────────
     All checks passed!